refactor: 기물 이동 규칙에 따른 경로 반환 클래스의 인터페이스, 클래스명 변경
- 기물 이동 규칙을 검증하는 방법은 두가지임: 이동 가능한 경로, 경로 도중에 장애물이 있는지 여부
- 각각 검증을 위한 인터페이스를 개별적으로 만들면서 '규칙'의 의미가 여러 클래스로 흩어지게 되었므로, 각 클래스가 검증하는 대상에 집중한 이름으로 네이밍 변경
