+++ src/main/java/janggi/domain/direction/RouteProvider.java
@@ -3,7 +3,7 @@ package janggi.domain.direction;
 import janggi.domain.Location;
 import java.util.List;
 
-public interface RouteRule {
+public interface RouteProvider {
 
     List<Location> calculateRoute(Location from, Location to);
 }
