+++ src/main/java/janggi/domain/piece/Gung.java
@@ -2,14 +2,14 @@ package janggi.domain.piece;
 
 import janggi.domain.Location;
 import janggi.domain.Side;
-import janggi.domain.direction.RouteRule;
-import janggi.domain.direction.GungSeongRouteRule;
+import janggi.domain.direction.RouteProvider;
+import janggi.domain.direction.GungSeongRouteProvider;
 import java.util.List;
 
 public class Gung extends Piece {
 
     private static final String PIECE_NAME = "궁";
-    private static final RouteRule ROUTE_RULE = new GungSeongRouteRule();
+    private static final RouteProvider ROUTE_PROVIDER = new GungSeongRouteProvider();
 
     public Gung(Side side) {
         super(PIECE_NAME, side);
@@ -23,7 +23,7 @@ public class Gung extends Piece {
     @Override
     public List<Location> calculateRoute(Location from, Location to) {
         try {
-            return ROUTE_RULE.calculateRoute(from, to);
+            return ROUTE_PROVIDER.calculateRoute(from, to);
         } catch (IllegalArgumentException e) {
             throw new IllegalArgumentException("궁은 해당 위치에 도달할 수 없습니다.");
         }
